Refactor: Moved Capturing() to helper module
