+++ modules/helpers/helper.py
@@ -1,0 +1,24 @@
+import sys
+from io import StringIO
+
+
+class Capturing(list):
+    """Capture stdout into a single buffer list.
+
+    >>> def func():
+            print('Hello')
+            print('World')
+    >>> with Capturing() as result:
+            func()
+    >>> result
+    ['Hello', 'World']"""
+
+    def __enter__(self):
+        self._stdout = sys.stdout
+        sys.stdout = self._stringio = StringIO()
+        return self
+
+    def __exit__(self, *args):
+        self.extend(self._stringio.getvalue().splitlines())
+        del self._stringio  # free up some memory
+        sys.stdout = self._stdout
